Moved timeline into old history list page
