+++ src/components/incidents-module/incidents-controller.js
@@ -81,7 +81,6 @@ class IncidentsController extends connect(store)(BaseController) {
         <impact-controller name="impact" route="{{subRoute}}"></impact-controller>
         <incident-comments name="comments"></incident-comments>
         <incident-history-controller name="history" route="{{route}}"></incident-history-controller>
-        <incident-timeline name="timeline"></incident-timeline>
       </iron-pages>
     `;
   }
@@ -173,10 +172,6 @@ class IncidentsController extends connect(store)(BaseController) {
         name: 'comments',
         tabLabel: 'COMMENTS',
         hidden: hideComments
-      },
-      {
-        name: 'timeline',
-        tabLabel: 'TIMELINE'
       },
       {
         name: 'history',
